+++ lib/glimmer/tk/text_proxy.rb
@@ -20,7 +20,6 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 require 'glimmer/tk/widget_proxy'
-require 'glimmer/tk/scrollable'
 
 module Glimmer
   module Tk
@@ -28,8 +27,6 @@ module Glimmer
     #
     # Follows the Proxy Design Pattern
     class TextProxy < WidgetProxy
-      include Scrollable
-      
       ALL_TAG = '__all__'
       FORMAT_DEFAULT_MAP = {
         'justify' => 'left',